arm: provide 64-bit switch for SUN50I
The Allwinner A64 SoC has 64-bit capable Cortex-A53 cores, but U-Boot
runs in AArch32 state. To launch an arm64 Linux kernel, we have to
call into ARM Trusted Firmware (which runs in AArch64) and ask it to
drop back into non-secure world in AArch64 state.
Allwinner has added a non-standard service call to achieve this.
Execution continues at the provided kernel entry point in r1, with
x0 containing the value we give it in r2.