sunxi: only write to CNTFRQ if really necessary
The ARM generic timer describes a register which holds the frequency
the generic timer is clocked with.
This register is writable only from the secure world, which we happen
to run in so far for most sunxi boards.
However the firmware stack on the Allwinner A64 is a bit different, so
we run U-Boot in non-secure state, which lets the mentioned register
write crash (U-Boot hangs).
As the firmware code running before U-Boot already set up the
register in our case, let's just read it and only try to write if the
value in there is wrong (or not set at all).
This way we still catch a non-populated register, while allow U-Boot
to easily run non-secure without further configuration.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>