sunxi: SPL: use best low level stack location for each SoC
Compiling the SPL in aarch64 for the Allwinner A64 yields a bigger code
size, which lets the lowlevel stack run into the end of the code.
Now with CONFIG_SPL_STACK properly honoured, let's move the initial stack
into the SRAM A2 on the A64.
The other SoCs stay at the end of SRAM A1, but the A80 can use the full
40K of it.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>